hwmon: Add driver for msm dal adc interface
This driver provides both a hwmon sysfs interface and ioctl
interface to the msm dal adc. The driver abstracts the multiple
adc devices found on the modem and presents them as a flat set
of channels. A lookup ioctl call can be made to discern which
local channel number corresponds to a particular name. The driver
supports both blocking and non-blocking conversion requests,
although only the blocking can be accessed from sysfs.

In non-blocking mode, a user may choose to block on a transfer
completion. The user may also choose to block on the availability
of transfer request slots, since this resource is finite.

Change-Id: Idcabd470632e6f7d032bebc83ede8f8a470b19b2
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>