thermal: add pmic8058 thermal manager driver
Add pmic8058-tm driver to monitor the temperature of PM8058
chip. There are 4 stages of temperature monitoring.
- stage 0: normal operation
- stage 1: temperature elevated
- stage 2: over temperature
- stage 3: critical temperature

When the temperature changes from one stage to another, an interrupt
will be generated. This event will be reported by the driver to the
operating system for thermal mitigation.

Stage 1 threshold can be programmed to one of 4 temperatures: 105, 110,
115 and 120 degree Celsius. Higher stages are 20 degrees(C) apart.

Change-Id: If08908932f57f14a42af66fd084eaddfdde800c6
Signed-off-by: Willie Ruan <wruan@codeaurora.org>
Signed-off-by: David Collins <collinsd@codeaurora.org>