msm: pil: Add peripheral reset ops
In the future there will be trusted or secure peripheral reset
functions. Currently, untrusted resetting is done directly in
the kernel using the peripheral reset driver. Trusted resetting
will be done by calling into the PAS (peripheral authentication
service) once that driver exists. For now, introduce the
operations so that in the future both implementations of the
reset code can coexist.

Change-Id: Ic7dc7c1cc0c13dd0f13bba3482d99de922b39d07
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>