mmc: core: reduce the data timeout values for CMD19 during the bustest.
The R2W factor doesnt really apply for CMD19 as there are no writes
to flash core for this command. Hence reduce the timeout value
by the R2W factor for CMD19. This has an advantage of reducing the
time it takes to complete the bustest by more than 2 seconds and hence
speeds up the card detection process.

Change-Id: I7792f71d1948c52b4ff74ee76cd2cf110c20aa18
Signed-off-by: Aparna Mallavarapu <aparnam@codeaurora.org>
[sthumma@codeaurora.org: Porting missing changes from .32 to .35]
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>