msm: 8x60: Migrate to timer based delay
A timer based udelay doesn't exhibit the problems that a jiffies
based udelay does with cpu frequency scaling. This is due to the
simple fact that the timer is running at a fixed rate unaffected
by cpu frequency scaling. Since this problem is more common on
SMP systems fix udelay() in those cases only. Also drop the
jiffies mucking that we do in the acpuclock driver since
loops_per_jiffy no longer has any correlation to the CPU
frequency.

Change-Id: Icf4113d5e1b26d89c5a390a711755e3f8de350f4
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>