msm: vidc: Enable/disable video encoder/decoder power rail as needed
On newer SoC like 8x60, power to the rotator hardware must be turned
off and on explicity.  Introduce use of 'regulator' APIs to turn on
and off power with the clocks.

Previous SoC that automatically turned power on and off as part of
clk_enable() and clk_disable() calls are still supported.  The
vidc driver assumes that if the call to regulator_get() is
unsuccessful, then the rail will be controlled automatically.

Change-Id: I8c0186dbd5c739c8ccadc2984684cfb5892c0e1c
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>