msm: clock-7x30: Fix PLL voting for disable case
During the clock merge we rewrote pll_enable() and pll_disable()
as one function that's multiplexed for the on and off case. The
off case wasn't ever written though, leaving only the pll_enable()
code in its place. Thus, we are never turning off PLL's when a
local_src_disable() call is made on 7x30. Fix it.

Change-Id: Iaefe8477d8ee3efce6c6712eaf7e70ace641491f
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>