mach-msm: sdio_al enhance wake up
SDIO-AL needs to toggle the pwrsave bit in order to properly
wake up the other side as part of the wakeup sequences of both
sides, and check via a GPIO if the other side has succefully woken up.
Otherwise, there are cases when the other side doesn't
wake up properly and all further operations of the wakeup
sequence fail.

Change-Id: I5e58bfca8ff5fa44900cf62d604f68a9bd7c50a2
Signed-off-by: Lena Salman <esalman@codeaurora.org>