msm: 8x60: usb_otg: Correct USB OTG regulator usage
USB OTG is attempting to set the voltage of regulator 8058_l6 with
a value that is outside of the specified constraint range.  Also,
it is enabling regulator 8058_s1 which it should not be touching.
Correct the regulator usage so that the system boots under the more
strict constraint checking included along with aggregation.

Change-Id: I3ddcd01109f37418818b6204c08af547ef8692a5
Signed-off-by: David Collins <collinsd@codeaurora.org>