msm: pil: Enable PLL4 before booting Q6
The Q6 needs PLL4 to be on before getting brought out of reset.
This would usually be handled by the driver requesting the
peripheral be brought out of reset, but with dependency chains we
cannot do that (the modem depends on the q6). Therefore, enable
the PLL before bringing it out of reset.

Change-Id: I9e74062f22082baa2c5eec01c1b95afcc5f58bcb
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>