defconfig: 8660: Enable CONFIG_MMC_UNSAFE_RESUME
If MMC_UNSAFE_RESUME is not defined, we enable the mmc bus suspend and
resume functions where we deselect the cards on suspend and re-detect
them on resume. With this config option enabled card re-initializations
can be prevented on every suspend resume cycle. However, it is assumed
that the card will not be physically removed and replaced with another
card while the system is in suspend state.

Change-Id: I21a9f6d2681a142fc3b60829f034e7c854f4cd11
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>