msm: 8660: audio: Fix I2C Problem by Changing PMGPIO Drive Strength
When the drive strength of the PMIC 8058 GPIOs 59 and 60 is 8mA,
then writing to the Timpani codec through the i2c bus causes i2c
 errors where Timpani drives the i2c data line low.  The errors
are "Transaction timed out" errors, with the status flags saying
that the I2C master no longer believes it is the master of the i2c
bus.  Once this error occurs, then every subsequent write or read
to Timpani fails, until Timpani is reset.

This error only appears after a large amount of Timpani register
writes, such as during a large number of audio device enables and
disables.  The problem also only appears on certain 8660 SURFs and
Timpani cards.

Whenever the drive strength is another value besides 8mA, such as
2mA, 4mA, or 12mA, then the error does not occur.  This patch sets
the drive strength to 12mA.  The i2c hardware team says that they
need to fix this problem so that a 8mA drive strength doesn't fail,
but this patch is a temporary solution to allow further audio
testing.

Change-Id: Ie1e6bffb2fb56819089c3f68ae941c6a39a2f054
Signed-off-by: Brad Rubin <brubin@codeaurora.org>