msm: msm_xo: Initialize XO state at boot
The XOs can be in an unknown state at boot so it's best to just
initialize their active and sleep sets to sane values to avoid any
issues in the future. Since the sleep set inherits the active set
by default, we don't set the sleep set for CXO buffers.

Change-Id: I13c7b0d355647d1a739c726c7b1080f6167bb8d7
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>