mmc: msm_sdcc: Prevent system suspend when SDIO request is pending
There is a chance where we get SDIO interrupt when function driver
suspend is completed and SDCC host suspend is in progress.  Prevent
system suspend in this case by acquiring a wakelock.  Otherwise, the
function driver resume will not be called to complete that request
and this results in timeout error for an application such as ping.

Change-Id: Iebf23d61747a4b1252a5c4fe91e01b54e881121a
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>