mmc: core: sdio: recognize io card without powercycle
Due to the bug in msm sdcc driver, call to sdio_reset() function
(newly added in 2.6.35 kernel) was causing bootup issues with
eMMC based targets (such as 8660) so we had bypassed the call to
sdio_reset() function as a temporary fix. Now as the actual
bug in msm sdcc driver is fixed, sdio_reset() function call
is again added back into mmc_rescan() function.

Below is the commit message from the actual patch
incorporated in 2.6.35 vanilla mainline linux.

SDIO Simplified Specification V2.00 states that it is
strongly recommended that the host executes either a power
reset or issues a CMD52 (I/O Reset) to re-initialize an I/O
only card or the I/O portion of a combo card. Additionally,
the CMD52 must be issued first because it cannot be issued
after a CMD0.

With this patch the Nintendo Wii SDIO-based WLAN card is
detected after a system reset, without requiring a complete
system powercycle.

Change-Id: I99e73461ec635b5a2abce41fb53b296d2d3e2a1f
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>