diag: Use current->comm to get current diag client name.
Use current->comm instead of get_task_comm(). get_task_comm() is not
exported function in kernel. to use diagchar as loadable module this
change is required. According to Linus, "current->comm" is fine,
and not racy. It only gets racy when you ask for the name of another
task.
Link:
http://www.collectedmail.com/archive/linux-kernel@vger.kernel.org/
				2009-01/Qco01A8GDjsudgEVcXhgsTne6hU.msg

Change-Id: I489816a3d0f04819766f866d05197c0c077386e4
Signed-off-by: Mayank Rana <mrana@codeaurora.org>