input: evdev: fix wakelock corruption
Corruption was observed with input subsystem's wakelock. The issue
was the data in the wakelock was zeroed out leading to a null pointer
access.The root cause was that the open function could release
the client structure without destroying the wakelock.

Call wake_lock_destroy in the failure path of the evdev_open()

Change-Id: I4a7a6be00485333c4e723f35c6da97a435fe22cf
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>