msm: saw-regulator: Add slew rate delay to set_voltage
The Saw regulator driver controls the Vdd supplied to the Scorpion
cores.  The PMIC 8901 automatically steps the output of the
regulators up slowly to avoid overshoot.  The slew rate is approx.
1.25 mV/us.

Add a call in the saw_set_voltage callback function to delay long
enough for this hardware process to finish.  The delay time is
proportional to the change in output voltage.

Change-Id: I1576ba234c17475ea0e1229b2a850b06e4eaf40f
Signed-off-by: David Collins <collinsd@codeaurora.org>