msm: timer: Cleanup global timer reading
The timer code is reading the CPU0 copy of the timer registers to
emulate a global timer. The address for CPU0's copy of the
registers is 0x40000 away from the base of the timer region and
thus the iomap is mapping 1MB of memory to get up to the CPU0
address with a simple addition. This wastes space in the iomap
and actually double maps regions in the virtual space. This isn't
very nice since we'd like to trap invalid accesses to reserved
regions. Therefore split the timer region into base and CPU0
parts and then do some maths to get the correct address in the
two places where the global timer is read.

Change-Id: I45bc6e5cd51402b8e47ffc7dca49accbe40d24c3
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>