msm: msm_sdcc: Reset SDCC in case of command timeout or CRC errors
Reset the controller for any errors to data command which does transfer
in non DMA mode. This ensures that data is flushed out from FIFOs.
Otherwise, there is a chance for any PIO interrupt to come after the
request is processed and this leads access to uninitialized scatter
gather list.

CRs-fixed: 257154
Change-Id: Ie157dbcf88314c0c4f2f08ead4165a74c7b2d767
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>