defconfig: syntax fix
The defconfig files are parsed both by make and the shell.  We use a
variable expansion to get the value of KERNEL_LOCAL_VERSION, but use
the make-specific syntax of $() to access it.  There doesn't appear to
be any use of this variable anymore.  Fix this by just removing the
expansion entirely.  The "perf" configs are annotated to distinguish
them from non-perf.

This eliminates the somewhat confusing message during build of
"KERNEL_LOCAL_VERSION: command not found", caused by the shell
interpreting the $() as a command substitution.

Change-Id: Ifc5be994a701d46616dc4a9e1537ad96e5dcd80a
Signed-off-by: David Brown <davidb@codeaurora.org>