msm: 8660: explicitly setting the direction for GPIO_SDC3_WP
For SDC3 slot, GPIO_SDC3_WP pin is used as input pin to detect
the position of write protect tab on SD card. By default after
power on this pin is configured as input but still it's not
safe to assume that it would be in input mode before
reading the status of the pin. With this patch, GPIO_SDC3_WP
is explicitly configured as input pin before reading the
status of the pin.

Change-Id: Icf46a7be0419604806b3451954830d2aabb7ca50
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>