msm: kgsl: Replace dev_priv list with a process_priv list
There isn't any reason to maintain a list of dev private structures
when the object we really care about are the per-process private
structures.  As a bonus, the per process list is easier to maintain
and lock at the kgsl_driver level.  Finally, change the inaccurate
kgsl_file_private structure name to kgsl_process_private.

Change-Id: I88d1906ae222ef66777f50096aebbfcfee9dc4a6
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>