msm: board-msm8x60: Regulator support for ffa/surf display
Previously the display code assumed that its regulator was left on, and
it didn't do any runtime configuring of it. Instead, turn its
regulator off when the display is not in use. This also has the
benefit of allowing the fluid display to run at a different voltage than
the surf/ffa display.

Change-Id: I1e3c0ba4d326b39f1a641d3b72af607d2cf35873
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>