msm: clock-8x60: Move gfx3d_clk hack from clock driver to board files
The gfx3d_clk on 8x60 cannot currently be gated or have its rate
changed without risking stability. A hack to keep the clock on
was previously hidden within clock-8x60.c.  Move it into the board
files for greater flexibility and visibility.

This allows enforcement of the requirement that the min and max
gfx3d_clk rates be the same and proper management of clk API
reference counts.

Change-Id: If7f2a43bc372a6f695c8ea6d5323818c49fb98eb
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>