msm: 8660: audio: Change d1 gpio to push/pull
Changes d1 gpio from pull-up and open drain to
push/pull.  This is appropriate because only
Timpani controls this gpio, and this will save
power by keeping the line low when it is not
in use. The d1 gpio controls the external
power amplifier used for some audio devices.
The d1 pin is initialized to 0 at startup.

Change-Id: I87408245a130ee001a45f28e8b2b765273b37ff7
Signed-off-by: Brad Rubin <brubin@codeaurora.org>