mmc: msm_sdcc: Always configure SDC CMD line in PUSH PULL mode
Currently during the card identification stage, SDC CMD line
is configured in OPEN DRAIN bus mode and after that it's configured
in PUSH PULL mode only during the transfer state (of card).
However It is recommended to use the PUSH PULL mode always
so that we can prevent signal degradation issues.
So this change ignores the bus_mode parameter received from
mmc core layer so that bus_mode is always set
to PUSH PULL mode.

Change-Id: I5dfd291a34c84738de8cd818560032d85bd257fe
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>