regulator: rpm-regulator: Use peak current voting instead of force mode
Change the rpm-regulator driver to send requests with an ip (peak
current) corresponding to high power or low power requirements.  The
driver currently sets a force mode value based on regulator consumer
needs. It is better to not specify a force mode because doing so
overrides aggregation on the RPM.

Change-Id: I7df15ea24c6bd716606e4cf34ff950c8b43c8f60
Signed-off-by: David Collins <collinsd@codeaurora.org>