msm: board-8x60: Change 8901_s4 regulator voltage constraints
8x60 is currently setting 8901_s4 to 1.3V during boot because that
regulator has a voltage constraint range of 1.3 - 1.3V and
apply_uV = 1 for it.  This works fine on a v1 power grid device as
Bluetooth, WLAN, and GPS need 1.3V from that regulator.  Unfortunately
8901_s4 switches purposes between the v1 and v2 power grid.  In the
v2 power grid, it is used by LPDDR2 memory that cannot be run at 1.3V.
Setting 8901_s4 to 1.3V has resulted in hardware damage on v2
devices.

Change the voltage constraints for 8901_s4 to be 1.225 - 1.225V and
set apply_uV = 0 for it so that it's voltage isn't set during regulator
registration.

This change is expected to break Bluetooth, WLAN, and GPS
functionality on v1 power grid devices.

Change-Id: Ie9bc91badd9d16b6fb8099d97da001c4c1a7b8b9
Signed-off-by: David Collins <collinsd@codeaurora.org>