msm: board 8x60: remove the use of 8901_s4 for BT
The load on 8901_s4 power rail in old power grid is swapped with the
load on 8058_s2 in new power rail. Since the detection of the grid
change is not provided to the kernel yet, temporarily remove the use
of 8901_s4 here to avoid incorrect voltage is set to the load in
new power grid.

Since boot code has set up 8901_s4 on old power grid and 8058_s2 on
new power grid correct, BT should work in both old and new power grid.
Power management has to wait until the grid info is available.

Change-Id: Id48b1e6187db3a23bcbade20447f272c83ab01d4

Signed-off-by: Willie Ruan <wruan@codeaurora.org>