msm: 8660: audio: Toggle Timpani sys_reset when doing hard reset
Toggles the Timpani sys reset line when doing a hard reset
on Timpani. Pulls the reset line high, then low, then high,
forcing Timpani to fully reset.

A 1ms busy delay between changing the reset state is
required based on hardware recommendations. A maximum
delay of 1.05 ms is good because it limits the Timpani
powerup time.

When the reset line is high, Timpani is on.

This fixes an issue where Tx audio (and sometimes Rx
audio) sparingly doesn't work.

Change-Id: I01282272669ba8e0168b69b4f8f7c7b16f712899
Signed-off-by: Brad Rubin <brubin@codeaurora.org>