mmc: msm_sdcc: Enable SDCC IRQ only when host is powered on
During idle-power collapse, MPM (Modem power management block)
will monitor all the enabled IRQs.  If SDCC IRQ is enabled, it
causes MPM to monitor the DATA1 GPIO line with level-low trigger.
This prevents TCXO shutdown during idle-power collapse if DATA1
is pulled low. Thus, disable SDCC IRQ when there is no device or
when the device is idle and is not a candidate for wakeup.

CRs-fixed: 261671
Change-Id: Iee250cf88cb52c46c7ca344486404407574c49e2
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>