msm: gpiomux: remove GPIOMUX_VALID.
GPIOMUX_VALID pawns complexity of implementation off on the
public interface, making the entire gpiomux library harder to use
in order to make gpiomux easier to write.  This is unacceptable.
Remove GPIOMUX_VALID, simplify the API, and absorb the complexity
of determining which gpiomux configs are used and which are not back
into the gpiomux implementation where it belongs.

Change-Id: I83e342c6e895a050c08e8e859369e4dff2e11296
Signed-off-by: Gregory Bean <gbean@codeaurora.org>