mmc: msm_sdcc: Fix race between msmsdcc_irq and set_ios methods
The variable host->clks_on needs protection with host lock anytime
it is accessed. Otherwise, it could lead to a scenario where SDIO
async IRQ occurs as soon as clocks are disabled but before the clks_on
is updated. Thus, the irq handler may not see the updated clks_on value
resulting in race condition.

Change-Id: Ifa0c2aa7e8aa9364b747e14501f8a55ab8c4f754
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>