msm: 8650a: Change pmem to ~4MB
pmem was earlier reduced to ~2MB because ashmem
is now used. However, the video player holds
one buffer when the device is put into standby.
While resuming the device it tries to acquire an extra
pmem buffer. This causes gralloc to run out of memory
for certain high resolution clips and thus show a
monochrome screen for a few moments until the buffer is
released. Hence, increasing the pmem size fixes the issue.

Change-Id: I72124fd2d6406da0d9129981927c3d9dc944801c
CRs-Fixed: 258977
Signed-off-by: Naseer Ahmed <naseer@codeaurora.org>