USB: msm_otg: Enable PHY comparators in LPM with wall charger
PMIC notifications are designed to receive VBUS high event only. We
should rely on PHY comparators for VBUS low interrupt. This does
not matter if we are connected to host. Because bus suspend is
not implemented. But if we don't enable PHY comparators and allow
LPM when wall charger is connected, we will not detect charger
disconnection.

Change-Id: I511babdb8928dda41f60f8cf67faffd7dd2d61b2
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>