arm: convert unofficial ARM machine type numbers to official ones
Several MSM platforms were using unofficial machine type
numbers rather than the official ones which are assigned.
The bootloaders are being converted to generate the
official machine type numbers, but the kernel needs to
recognize and convert the unofficial ones during a
transition period.

The patch also brings our copy of mach-types up to date
with the entire official list.

Change-Id: I2b3e1cb324cbdb715b3b4043d68970a670c885ed
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>