Revert "msm: acpuclock-7x30: Vote for msmc1"
This reverts commit 847c0fd777d31640f7bcf3266117ba4057248e33.

The reverted commit was being explicit about increasing MSMC1
when the cpu frequency is high. Before this commit, the MSMC1
level was implicity increased via EBI1 rate requests. The current
approach (explicit voting) doesn't allow MSMC1 to drop when the
cpu is in power collapse. Therefore, revert the explicit voting
and rely on the previous behavior where MSMC1 is voted on
implicitly. This allows the modem to drop the MSMC1 vote from the
acpu clock driver during power collapse at high frequencies.

Change-Id: I28e2da61cf67bc52f43614b04fdd6b4df4616fb5
CRs-Fixed: 264047
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>