msm: pil: Don't request empty blobs
If a blob is of size 0 on disk we'll request an empty blob from
userspace and then fail to find the blob. Since
request_firmware() specifically returns -ENONENT when the
firmware size is 0 we can't just have empty blobs on disk.
Instead, skip requesting blobs for segments that have no on disk
size but still zero out any memory for the segment.

Change-Id: I77ca37258060427a1d6a45f7b11af86a39693b22
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>