msm: kgsl: use dma coherent memory for shared memory regions
The use of ioremap() in kgsl_sharedmem.c is incorrect and
soon to be deprecated. See:
http://thread.gmane.org/gmane.linux.kernel/1045670

This is the first of a series of patches to wean kgsl off of
this memory region.

This patch switches all in kernel allocations, except for
context state, to use memory from dma_alloc_coherent().
These allocations are done at probe time, with the
exception of context state and mmu pagetables when
KGSL_PER_PROCESS_PAGE_TABLE is enabled.

Context state will be moved to paged memory in a separate
patch.

KGSL_PER_PROCESS_PAGE_TABLE, which is off by default on all
targets, is likely to not function correctly with this patch,
because when this option is enabled we're counting on being
able to allocate ~32 contiguous pages at runtime which is
not likely to succeed. This will be addressed by creating a
pool of pagetables at probe time in a separate patch.

Change-Id: Ia4d0558c6cbe7340488c8b9fefbbfd864ee454f6
CRs-fixed: 261204
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>