msm: board-msm8x60: hdmi: Request new 8901_mpp0 regulator
The 8901_mpp0 "regulator" acts as a switch to control the external
5V regulator which supplies the 8901_hdmi_mvs switch.  Enable this
regulator when HDMI 5V is turned on.

Also, add checks so that regulator_enable and regulator_disable
are not called multiple times in a row because this messes up the
regulator framework's reference count for the given regulator.

Remove unnecessary private 8901_mpp0 initialization.

Change-Id: I65d1b6fe1e0ccb2c338bd129dac4d1856889ab9a
Signed-off-by: David Collins <collinsd@codeaurora.org>