mmc: core: Remove machine macro usage from SDIO driver
The Qualcomm SDIO client on MDM contains NULL CIS tuple for all
of its SDIO functions.  As a temporary software workaround
the machine_is_xxxx macros are curently being used to skip the
reading of CIS.  But it would also effect other SDIO cards used
on that target which is wrong.  Thus, this patch removes the usage
of machine macros in SDIO driver and checks for the vendor-id and
device-id of SDIO client on MDM to skip CIS register read.

CRs-fixed: 266272
Change-Id: I819a056e7d72e7f0e401454a89623cbc652e756d
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>