msm: watchdog: Add a watchdog bark interrupt handler
Add a bark interrupt handler in an attempt to make watchdog
bite events more visible in the debug logs. Set the bark
timeout equal to the current bite timeout, and increase the
bite timeout by one second to give time for the bark
handler to run. Add a panic message to the bark handler to
indicate that the panic was caused by the watchdog, and
print the last pet time as well as the current time.
The side effect is a one-second increase in the latency of
a watchdog reset if the watchdog bark occurs when
interrupts are disabled, as well as a potentially longer
latency for a watchdog reset with interrupts enabled, since
the reset will be delayed until the panic logic had run and
the panic timeout has expired.

Not all watchdog bark events will be registered, due to the
possiblity that the bark can happen with interrupts
disabled or when the chip is otherwise wedged. But, this
may nevertheless be useful for debugging.

Change-Id: I22a13be832559e63b6c5c47d6112f71402ccf0f5
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>