msm: timer: DGT fixes for suspend
Suspend is broken when the DGT is the global timer because it is
left on for the purposes of udelay. Being left on causes it to generate
spurious interrupts which keep the cpu from power collapsing. So, only
keep cpu 0's dgt on, and set its match value to 0 when we disable its
clockevent

Change-Id: Ie9ce0b995db671d184aab2c37c3f6a4e9fe3dc2e
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>