msm: qdsp5v2: Reorganize files to export apis to ALSA framework
ALSA VOIP driver need to switch from modem to QDSP6, the api to
do this is implemented in voice proxy driver. Which needs to be
exported so that ALSA framework can use it. Hence the header file
is moved to standard include path.

For the new mixer control for FTM loopback need afe apis. Hence
the afe header file is moved to standard include path.

Change-Id: Ia91aac5911a155a48e560ecf0c97518798d337ca
Signed-off-by: Deepa Madiregama <dmadireg@codeaurora.org>