msm: 8660: configure pmic8058 as gpio direct interrupt
The gpio tlmm summary line is used to notify the main interrupt
controller about pmic interrupts. The summary line isnt operational when
the XO clock is turned off.

Move the pmic interrupt to be a direct interrupt rather than going over
the sumamry line.

Note that pmic is a active low interrupt and the main interrupt
controller(GIC) only supports active high interrupt. Make the
gpio block change the polarity, such that it detects a active
low interrupt but generate a active high interrupt
to the GIC.

Change-Id: I4a9901b32d50dd75a834456cfea30b2cf08a28c9
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>