msm: rpm-regulator: Fix save_uV value for sleep-selectable regulators
The save_uV variable is used to store the voltage of a regulator in
the rpm-regulator driver.  For sleep-selectable regulators which are
accessible from the private set_voltage API, save_uV is not set
correctly.  The value passed in from the consumer driver is used
directly instead of the internally aggregated value.

Change the rpm-regulator driver so that the aggregated voltage is
stored in save_uV for sleep-selectable regulators.

Change-Id: I1c3b7aa4dee666d904b96872ed632c2e91e7d6bf
Signed-off-by: David Collins <collinsd@codeaurora.org>