msm: rpc: Maintain and use a pool of small buffers
In the read direction, a huge 16 KB memory is initially allocated.
But only a small fraction of it is getting used. In the write direction,
buffers are allocated for the data to be transmitted using GFP_KERNEL
mask. But the RPC Router is holding spinlock while performing write.
This sometimes leads to memory allocation failure and hence write
failure. In either ways, maximum message size is 536. So preallocate,
maintain and use a pool of small buffers for read and write.

Change-Id: I3fea4f5cf1dc3d7342ddd6e42ead20f89cc288c7
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>