msm: cpufreq: Fix race condition in cpufreq with suspend
While the CPU is in the process of changing the frequency, it is
posisble that the system can go into suspend. Suspend directly
calls acpuclk_set_rate() to set the frequency, which messes up the
internal PLL information as frequency change already initated by
CPUFreq is in progress. Hold off the suspend process until the
frequency change is complete.

Change-Id: Ief07ccfbe3d402479e39164f4e94e4b958678929
Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>