msm: Add flashless boot support to sdio_al
When the MDM SDIOC is in boot mode it writes a unique signature
in the SW mailbox header. When sdio_al identified this signature
it activates the sdio_al bootloader and waits for its completion.
Upon its completion it reads the SW mailbox again to synchronize
with the SDIOC configuration.

Change-Id: I9fd0aa24ec3c2cd491f263e04215b33873c4f93a
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>