ASoC: 8x60: Add device switch support for PCM sessions.
The PCM session is to be used for playback/capture and
needs to be routed to different sinks/sources. Add support to
switch dynamically between different sinks/sources.

Remove current assignment of the index to simple controls as they
are not relevant and should be used by the device controls to
identify the device number.

Add support for multisession for playback.

(cherry picked from commit d0f7d817d8da167cffb355524c90f1dfbdece1ee)

Change-Id: Ie096b14de67f56bf9a6f057366adb4ae21aafd42
Signed-off-by: Asish Bhattacharya <asishb@codeaurora.org>