pmic8058-charger: fix msm_xo_voting panic in pmic charger
A panic was observed when the pmic usb charger was disabled and the usb
suspends. USB suspend leads to a call to the pmic usb charger driver to
stop charging, which votes to turn off the D1 clock.

The voter may not be initialized while voting. Hence vote only if the
voter object was initialized.

Change-Id: I946e5b9225d6e084c0c6c1d2c815fdde759094d9
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>