diag: Pass correct parameters to overflow check macro
CHK_OVRFLOW takes as its arguments the start of the
buffer, the start of the location in the buffer where
data is to be copied, the end of the buffer, and the
length of the data to be copied. diag_update_log_masks
is passing a value greater than the maximum size of the
log masks buffer as the third argument to CHK_OVERFLOW,
fix this to be the maximum size. Also update the log mask
size to accommodate bigger mask requests.

Change-Id: I7b033b1654a161bde04541c6eccc826c28c0eb57
CRs-Fixed: 270142
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>