msm: board-msm8x60: Set peak_uA for RPM regulators to physical HPM values
Change the peak_uA values specified for each regulator controlled by the
rpm-regulator driver to match the physical HPM threshold value for the
regulator.  Previously all regulators were set to 51000 uA to match the
manner in which the RPM chose between LPM and HPM.

This change will still keep all regulators in HPM by default, but will
change the peak current value to the minimum value required to do so.

Change-Id: Iefedc3008439f16cdb84312fee52515a96cee93e
Signed-off-by: David Collins <collinsd@codeaurora.org>