PM: increase dpm timeout value to 10
The current dpm watchdog timer is set to 3 second. Under this value,
dpm watchdog timeout is observered when kgsl driver failed to be
suspended on time. As a workaround, the timeout value is set to 10.
This shall be adjusted when the root cause is identified.

Signed-off-by: Wei Zou <wzou@codeaurora.org>
(cherry picked from commit 6000aa2bb587916403cdd7ad1458a9d2d49b1e34)

Change-Id: I2a4c6c03eee7243cf39a65ae9c8418742cd87aa9
Signed-off-by: Dominic Binks <dbinks@codeaurora.org>