msm: 8x60: Remove "always_on" vote for 8901_L5 regulator
8901_L5 regulator powers eMMC card's VCC voltage. During system
suspend it is not safe to disable this regulator as the card may be
in unknown state during wakeup and card detection failure happens.
SDCC driver is now taking care of this by not voting to disable this
regulator but keep in low power mode with load current 9mA. Hence, remove
always on vote for 8901_L5 regulator and let the driver decide the mode of
operation.

Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>
(cherry picked from commit fac4f8551e2135cab35237178212dd93e88d1e16)

Change-Id: I3bad1e78693fe6ad28d621851c0fe85d20ddb4c8
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>